refactor: add useExperimentParameters
